requirements.txt
gspread>=3.6.0
requests>=2.20.0
e-objects>=1.2.1
GitPython==3.1.11
beautifulsoup4==4.8.0
google-api-python-client==1.10.0
